Rename spec_helper rails_helper to follow convention
