+++ spec/system/pandas/index_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 RSpec.describe 'home page', type: :system do
   context 'when navigating from the home page to the pandas index page' do
